refactor(api): move feature toggle serializer to src
