+++ api/src/shared/application/feature-toggles/feature-toggle-controller.js
@@ -1,5 +1,5 @@
-import * as serializer from '../../../../lib/infrastructure/serializers/jsonapi/feature-toggle-serializer.js';
 import { config } from '../../../../src/shared/config.js';
+import * as serializer from '../../infrastructure/serializers/jsonapi/feature-toggle-serializer.js';
 
 const getActiveFeatures = function () {
   return serializer.serialize(config.featureToggles);
