+++ api/tests/shared/unit/infrastructure/serializers/jsonapi/feature-toggle-serializer_test.js
@@ -1,5 +1,5 @@
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/feature-toggle-serializer.js';
-import { expect } from '../../../../test-helper.js';
+import * as serializer from '../../../../../../src/shared/infrastructure/serializers/jsonapi/feature-toggle-serializer.js';
+import { expect } from '../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | feature-toggle-serializer', function () {
   describe('#serialize', function () {
